+++ container/presentation-page/Hero.jsx
@@ -1,8 +1,8 @@
 import { Eyes } from "@/components";
+
 export default function Hero() {
 	return (
 		<section className="w-full padding-x bg-[#6495ED] sticky top-0 h-[65vh]">
-
 			<div className="w-full pt-[170px] pb-[20px]">
 				<div className="w-fit relative">
 					<h1 className="heading tracking-[-1.3px] text-[#212121] font-semibold font-FoundersGrotesk uppercase">
