+++ container/services-page/Capibilyties.jsx
@@ -7,9 +7,9 @@ import { serviceCapaybilitiesItem } from "@/constants";
 export default function Capibilyties() {
 	const [hovered, setHovered] = useState(false);
 	const [hovered1, setHovered1] = useState(false);
-	return (
-		<div className="w-full padding-y rounded-t-[20px]" style={{ backgroundColor: '#6495ED' }}>
 
+	return (
+		<div className="w-full padding-y rounded-t-[20px]" style={{ backgroundColor: "#6495ED" }}>
 			<div className="w-[82%] sm:w-full xm:w-full padding-x mb-[70px]">
 				<h1 className="sub-heading font-medium font-NeueMontreal text-secondry">
 					<span className="xl:link-flash lg:link-flash md:link-flash cursor-pointer">
@@ -24,17 +24,21 @@ export default function Capibilyties() {
 					</span>
 				</h1>
 			</div>
+
 			<div className="w-full flex sm:flex-col xm:flex-col justify-between py-[20px] padding-x border-t border-[#21212155] sm:gap-[15px] xm:gap-[15px]">
 				<div className="w-[10%] sm:w-full xm:w-full">
 					<h3 className="paragraph font-medium text-secondry font-NeueMontreal">
 						Our Capabilities:
 					</h3>
 				</div>
+
 				<div className="w-[90%] flex flex-col gap-y-[30px]">
 					{serviceCapaybilitiesItem.map((item) => (
 						<div
 							className="w-full flex sm:flex-col xm:flex-col justify-between sm:gap-[15px] xm:gap-[15px]"
-							key={item.id}>
+							key={item.id}
+						>
+							{/* Image on hover */}
 							<div className="w-[40%] flex justify-end mr-[150px] sm:hidden xm:hidden">
 								{hovered && item.id === 1 ? (
 									<Image
@@ -57,6 +61,8 @@ export default function Capibilyties() {
 									)
 								)}
 							</div>
+
+							{/* Left column */}
 							<div className="w-[25%] sm:w-full xm:w-full">
 								<div className="flex gap-x-[10px] flex-col">
 									<div className="flex items-center gap-x-[10px] pb-[20px]">
@@ -69,16 +75,15 @@ export default function Capibilyties() {
 								<div
 									className="flex flex-col gap-y-[5px]"
 									onMouseEnter={() => setHovered(item.id === 1 && true)}
-									onMouseLeave={() => setHovered(item.id === 1 && false)}>
+									onMouseLeave={() => setHovered(item.id === 1 && false)}
+								>
 									{item.links1.map((link) => (
-										<Button
-											href={link.href}
-											title={link.title}
-											key={link.id}
-										/>
+										<Button href={link.href} title={link.title} key={link.id} />
 									))}
 								</div>
 							</div>
+
+							{/* Right column */}
 							<div className="w-[25%] sm:w-full xm:w-full">
 								<div className="flex gap-x-[10px] flex-col">
 									<div className="flex items-center gap-x-[10px] pb-[20px]">
@@ -91,13 +96,10 @@ export default function Capibilyties() {
 								<div
 									className="flex flex-col gap-y-[5px]"
 									onMouseEnter={() => setHovered1(item.id === 1 && true)}
-									onMouseLeave={() => setHovered1(item.id === 1 && false)}>
+									onMouseLeave={() => setHovered1(item.id === 1 && false)}
+								>
 									{item.links2.map((link) => (
-										<Button
-											href={link.href}
-											title={link.title}
-											key={link.id}
-										/>
+										<Button href={link.href} title={link.title} key={link.id} />
 									))}
 								</div>
 							</div>
