+++ container/workiz-page/Video.jsx
@@ -1,5 +1,5 @@
 import { PlayVideo } from "@/components";
 
 export default function Video() {
-	return <PlayVideo videosrc="/workizvideo.mp4" />;
+    return <PlayVideo videosrc="/workizvideo.mp4" />;
 }
